Crash and restart on memory error
This is a band-aid fix for out of memory errors we are seeing from time
to time. With this configuration, the service that runs out of heap memory
will crash and Docker will restart it.

This a fix for a situation we are observing - a service runs out of memory
and then is stuck in a broken state.